+++ tests/basic.rs
@@ -5,8 +5,16 @@ mod basic {
     //
     #[test]
     fn size_of() {
-        assert_eq!(std::mem::size_of::<OptStr>(), 184);
-        assert_eq!(std::mem::size_of::<MetaType>(), 32);
+        #[cfg(target_pointer_width = "64")]
+        {
+            assert_eq!(std::mem::size_of::<OptStr>(), 184);
+            assert_eq!(std::mem::size_of::<MetaType>(), 32);
+        }
+        #[cfg(target_pointer_width = "32")]
+        {
+            assert_eq!(std::mem::size_of::<OptStr>(), 96);
+            assert_eq!(std::mem::size_of::<MetaType>(), 16);
+        }
         assert_eq!(std::mem::size_of::<Pasc>(), 1);
     }
 }